+++ app/components/notifications/styles/notifications.js
@@ -1,4 +1,4 @@
-import { flexContainer } from 'styles/variables';
+import { flexContainer, defaultFont } from 'styles/variables';
 
 const styles = (theme) => ({
   flexContainer: {
@@ -15,6 +15,7 @@ const styles = (theme) => ({
     width: '100%',
     borderTopLeftRadius: 0,
     borderTopRightRadius: 0,
+    padding: theme.spacing(1),
   },
   toolbar: {
     width: '100%',
@@ -43,6 +44,51 @@ const styles = (theme) => ({
   },
   hasFilterBlur: {
     filter: 'blur(15px)',
+  },
+  flex: {
+    ...flexContainer,
+    flexDirection: 'column',
+    justifyContent: 'flex-start',
+    alignItems: 'flex-start',
+  },
+  tableRow: {
+    border: 'none',
+    padding: theme.spacing(1),
+    lineHeight: '1.1',
+    verticalAlign: 'middle',
+    '&:hover': {
+      cursor: 'pointer',
+    },
+  },
+  tableCell: {
+    '& p': {
+      overflowWrap: 'break-word',
+    },
+  },
+  cellText: {
+    ...defaultFont,
+    textAlign: 'left',
+    whiteSpace: 'nowrap',
+    overflow: 'hidden',
+    textOverflow: 'ellipsis',
+  },
+  requiredBy: {
+    [theme.breakpoints.up('lg')]: {
+      width: 'auto',
+    },
+    [theme.breakpoints.down('md')]: {
+      width: 200,
+    },
+    whiteSpace: 'nowrap',
+    overflow: 'hidden',
+    textOverflow: 'ellipsis',
+    textAlign: 'left',
+  },
+  span: {
+    color: '#fff',
+  },
+  popover: {
+    pointerEvents: 'none',
   },
 });
